+++ plugins/O-IntonationPad/Source/DSP/EQProcessor.h
@@ -31,6 +31,8 @@ private:
     using Filter = juce::dsp::IIR::Filter<float>;
     using StereoFilter = juce::dsp::ProcessorDuplicator<Filter, FilterCoeffs>;
 
+    void updateCoefficients();
+
     StereoFilter lowShelf;
     StereoFilter midPeak;
     StereoFilter highShelf;
@@ -42,4 +44,10 @@ private:
     std::atomic<float> targetMidGainDB { 0.0f };
     std::atomic<float> targetMidFreqHz { 1000.0f };
     std::atomic<float> targetHighGainDB { 0.0f };
+
+    // Dirty-flag caching: only recompute coefficients when parameters change
+    float appliedLowGainDB = -999.0f;
+    float appliedMidGainDB = -999.0f;
+    float appliedMidFreqHz = -999.0f;
+    float appliedHighGainDB = -999.0f;
 };